+++ src/git/index.ts
@@ -1,3 +1,5 @@
+// Licensed under the Hungry Ghost Hive License. See LICENSE.
+
 export * from './branches.js';
 export * from './github.js';
 export * from './submodules.js';